# Fix for large period