* Fix an `InexactError` that occurred when `rationalify()`ing `RepeatingDecimal`s with large period
* Improve performance of `rationalify()`.