@hyrodium 2023-12-27
Thank you for the improvements!

>Improve performance of `rationalify()`.

**Before this PR**
```julia
julia> using RepeatingDecimalNotations, BenchmarkTools

julia> using RepeatingDecimalNotations: stringify, rationalify

julia> rationalify("123.456(052631578947368421)")
469133//3800

julia> @benchmark rationalify("123.456(052631578947368421)")
BenchmarkTools.Trial: 10000 samples with 8 evaluations.
 Range (min … max):  2.966 μs …  4.306 ms  ┊ GC (min … max):  0.00% … 60.33%
 Time  (median):     3.249 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   4.599 μs ± 69.474 μs  ┊ GC (mean ± σ):  15.42% ±  1.02%

     ▄▇█▇▆▃▁▁                                                 
  ▁▃▇██████████▅▅▄▄▅▅▅▅▄▄▃▃▃▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▂▂▂▂▂▂▂▁▁▁▁▁ ▃
  2.97 μs        Histogram: frequency by time        4.73 μs <

 Memory estimate: 1.77 KiB, allocs estimate: 57.
```

**After this PR**
```julia
julia> using RepeatingDecimalNotations, BenchmarkTools

julia> using RepeatingDecimalNotations: stringify, rationalify

julia> rationalify("123.456(052631578947368421)")
469133//3800

julia> @benchmark rationalify("123.456(052631578947368421)")
BenchmarkTools.Trial: 10000 samples with 9 evaluations.
 Range (min … max):  2.812 μs …  3.385 ms  ┊ GC (min … max):  0.00% … 62.89%
 Time  (median):     3.285 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   3.982 μs ± 47.359 μs  ┊ GC (mean ± σ):  10.03% ±  0.84%

    ▂▃       ▇█▃                                              
  ▂▇███▅▅▄▄▄█████▆▄▃▂▃▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  2.81 μs        Histogram: frequency by time        5.19 μs <

 Memory estimate: 1.43 KiB, allocs estimate: 47.
```